feat: BankManager의 visitClients 메서드 구현
