+++ BankManagerConsoleApp/BankManagerConsoleApp/BusinessType.swift
@@ -1,3 +1,4 @@
+
 enum BusinessType {
     case basic
 }
