+++ BankManagerConsoleApp/BankManagerConsoleApp/Client.swift
@@ -1,3 +1,4 @@
+
 struct Client {
     let waitingNumber: Int
     let business: BusinessType
