+++ BankManagerConsoleApp/BankManagerConsoleApp/main.swift
@@ -4,13 +4,25 @@
 //  Copyright © yagom academy. All rights reserved.
 // 
 
-import Foundation
+//var bankManager = BankManager()
+//private var selectedMenu = bankManager.selectMenu()
+//
+//while selectedMenu == 1 {
+//    print("은행개점")
+//    selectedMenu = bankManager.selectMenu()
+//}
+//print("종료")
 
-var bankManager = BankManager()
-private var selectedMenu = bankManager.selectMenu()
-
-while selectedMenu == 1 {
-    print("은행개점")
-    selectedMenu = bankManager.selectMenu()
+func selectMenu() -> Int {
+    print(" 1 : 은행개점 \n 2 : 종료\n 입력 :", terminator: " ")
+    
+    if let menuNumber = Int(readLine() ?? "") {
+        return menuNumber
+    }
+    
+    return 0
 }
-print("종료")
+
+//func printMenu() {
+//    
+//}
